+++ spectator/src/test/java/net/boreeas/riotapi/spectator/chunks/BlockStreamReaderTest.java
@@ -20,36 +20,59 @@ import junit.framework.TestCase;
 import lombok.extern.log4j.Log4j;
 import net.boreeas.riotapi.Shard;
 import net.boreeas.riotapi.spectator.GamePool;
+import net.boreeas.riotapi.spectator.GameUpdateTask;
 import net.boreeas.riotapi.spectator.InProgressGame;
 import net.boreeas.riotapi.spectator.SpectatorApiHandler;
 
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.Map;
+import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.atomic.AtomicInteger;
 
 @Log4j
 public class BlockStreamReaderTest extends TestCase {
 
+    private static CountDownLatch latch = new CountDownLatch(1);
     private static ArrayList<Block> blocks = new ArrayList<>();
 
     static {
         SpectatorApiHandler handler = new SpectatorApiHandler(Shard.EUW);
         InProgressGame game = handler.openFeaturedGame(handler.getFeaturedGames().get(0));
-        GamePool pool = GamePool.singleton(game);
+        GamePool pool = new GamePool();
+        GameUpdateTask task = pool.submit(game);
 
-        BlockStreamReader reader = new BlockStreamReader(game.getKeyFrame(game.getFirstAvailableKeyFrame()).getBuffer());
-        while (reader.hasNext()) {
-            blocks.add(reader.next());
-        }
+        task.setOnKeyframePulled(i ->
+            new Thread(() -> {
+                synchronized (BlockStreamReader.class) {
+                    if (latch.getCount() == 0) {
+                        return;
+                    }
+
+                    BlockStreamReader reader = new BlockStreamReader(game.getKeyFrame(i).getBuffer());
+                    while (reader.hasNext()) {
+                        blocks.add(reader.next());
+                    }
+
+                    latch.countDown();
+                    pool.shutdown();
+                }
+            }).start()
+        );
     }
 
     public void testNext() throws Exception {
+
+        try {
+            latch.await();
+        } catch (InterruptedException e) {
+            e.printStackTrace();
+        }
+
         assertFalse(blocks.isEmpty());
 
         Map<Integer, AtomicInteger> counts = new HashMap<>();
         for (Block block: blocks) {
-            logBlock(block);
             if (counts.containsKey(block.getHeader().getType())) {
                 counts.get(block.getHeader().getType()).incrementAndGet();
             } else {
@@ -59,14 +82,9 @@ public class BlockStreamReaderTest extends TestCase {
 
         log.info("\n===== Counts =====");
         for (Map.Entry<Integer, AtomicInteger> countEntry: counts.entrySet()) {
-            log.info(String.format("[%2x.d] %d", countEntry.getKey(), countEntry.getValue().get()));
+            log.info(String.format("[%2x] %d", countEntry.getKey(), countEntry.getValue().get()));
         }
         log.info("Total: " + blocks.size());
     }
 
-    private void logBlock(Block block) {
-        String fmt = "[%2x.d] t=%6.dms p=%x.d l=%d";
-        BlockHeader header = block.getHeader();
-        log.info(String.format(fmt, header.getType(), header.getTimestamp(), header.getTimestamp(), header.getBlockParam(), header.getContentLength()));
-    }
 }